devex-1197: fix imports after move
